changes store to utils
