+++ src/components/Preferences/FormStyles.js
@@ -1,4 +1,4 @@
-import { makeStyles, InputBase, withStyles } from '../../store/index';
+import { makeStyles, InputBase, withStyles } from '../../utils/index';
 
 const useStyles = makeStyles({
   root: {
