+++ src/utils/index.js
@@ -1,4 +1,5 @@
 export * from './icons';
 export * from './images/index';
+export * from './helpers/time-helpers';
 export * from './constants';
-export * from './material-components'
+export * from './material-components';
